Move model classes to separate package
