+++ src/main/java/me/chrr/scribble/mixin/BookEditScreenMixin.java
@@ -5,10 +5,10 @@ import it.unimi.dsi.fastutil.ints.IntArrayList;
 import it.unimi.dsi.fastutil.ints.IntList;
 import me.chrr.scribble.Scribble;
 import me.chrr.scribble.book.*;
-import me.chrr.scribble.book.command.BookEditScreenCutCommand;
-import me.chrr.scribble.book.command.BookEditScreenInsertCommand;
-import me.chrr.scribble.book.command.BookEditScreenMemento;
-import me.chrr.scribble.book.command.BookEditScreenPasteCommand;
+import me.chrr.scribble.model.command.BookEditScreenCutCommand;
+import me.chrr.scribble.model.command.BookEditScreenInsertCommand;
+import me.chrr.scribble.model.memento.BookEditScreenMemento;
+import me.chrr.scribble.model.command.BookEditScreenPasteCommand;
 import me.chrr.scribble.gui.ColorSwatchWidget;
 import me.chrr.scribble.gui.IconButtonWidget;
 import me.chrr.scribble.gui.ModifierButtonWidget;
