+++ src/main/java/me/chrr/scribble/model/memento/BookEditScreenMemento.java
@@ -1,22 +1,22 @@
-package me.chrr.scribble.book.command;
+package me.chrr.scribble.model.memento;
 
-import me.chrr.scribble.book.RichSelectionManager;
 import me.chrr.scribble.book.RichText;
 import me.chrr.scribble.tool.commandmanager.Command;
+import net.minecraft.client.gui.screen.ingame.BookEditScreen;
 import net.minecraft.util.Formatting;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Set;
 
 /**
- * Represents the mutable state of a {@link RichSelectionManager},
- * which can be modified through the execution of RichSelectionManager___Command.
+ * Represents the mutable state of a {@link BookEditScreen},
+ * which can be modified through the execution of BookEditScreen___Commands.
  * See {@link Command}.
  */
-public record RichSelectionManagerMemento(
+public record BookEditScreenMemento(
         int selectionStart,
         int selectionEnd,
-        RichText richText,
+        RichText currentPageRichText,
         @Nullable Formatting color,
         Set<Formatting> modifiers
 ) {
